SeoDash - Web

+++ Aplicaciones/Proposito/views.py
@@ -2,6 +2,7 @@ from django.shortcuts import render, redirect, get_object_or_404
 from django.contrib import messages
 from .models import Proposito
 from django.contrib.auth.decorators import login_required, user_passes_test
+from Aplicaciones.SesionMeditacion.models import SesionMeditacion  
 
 def es_admin(user):
     return user.is_staff or (hasattr(user, 'perfilusuario') and user.perfilusuario.rol.upper() in ['ADMIN', 'ADMINISTRADOR'])
@@ -81,6 +82,13 @@ def guardarEdicionProposito(request, id):
 @user_passes_test(es_admin)
 def eliminarProposito(request, id):
     proposito = get_object_or_404(Proposito, pk=id)
+
+    # Verifica si el propósito está relacionado con alguna sesión
+    if SesionMeditacion.objects.filter(proposito=proposito).exists():
+        messages.error(request, "No se puede eliminar el propósito porque está en uso en sesiones de meditación.")
+        return redirect('propositos')
+
+    # Si no está relacionado, se puede eliminar
     proposito.delete()
     messages.success(request, "Propósito eliminado correctamente")
     return redirect('propositos')
